add module docstrings
